+++ nucypher-core-wasm/tests/wasm.rs
@@ -672,7 +672,7 @@ fn metadata_response() {
 
 #[wasm_bindgen_test]
 fn request_public_key() {
-    let secret = RequestSecretKey::random();
+    let secret = SessionStaticSecret::random();
     let public_key = secret.public_key();
 
     assert_eq!(secret.public_key(), secret.public_key());
@@ -680,7 +680,7 @@ fn request_public_key() {
     // mimic transmission public key over the wire
     let serialized_public_key = public_key.to_bytes();
     let deserialized_public_key =
-        RequestPublicKey::from_bytes(serialized_public_key.as_ref()).unwrap();
+        SessionStaticKey::from_bytes(serialized_public_key.as_ref()).unwrap();
 
     assert_eq!(public_key, deserialized_public_key);
     assert_eq!(serialized_public_key, deserialized_public_key.to_bytes());
@@ -689,10 +689,10 @@ fn request_public_key() {
 #[wasm_bindgen_test]
 fn threshold_decryption_request() {
     let ritual_id: u32 = 5;
-    let service_secret = RequestSecretKey::random();
+    let service_secret = SessionStaticSecret::random();
     let service_public_key = service_secret.public_key();
 
-    let requester_secret = RequestSecretKey::random();
+    let requester_secret = SessionStaticSecret::random();
 
     let conditions = "{'some': 'condition'}";
     let conditions_js: JsValue = Some(Conditions::new(conditions)).into();
@@ -737,9 +737,8 @@ fn threshold_decryption_request() {
     assert_eq!(request, decrypted_request);
 
     // wrong key used
-    let random_secret_key = RequestSecretKey::random();
-    let random_shared_secret =
-        RequestSharedSecret::from(random_secret_key.derive_shared_secret(&service_public_key));
+    let random_secret_key = SessionStaticSecret::random();
+    let random_shared_secret = random_secret_key.derive_shared_secret(&service_public_key);
     assert!(encrypted_request_from_bytes
         .decrypt(&random_shared_secret)
         .is_err());
@@ -749,9 +748,9 @@ fn threshold_decryption_request() {
 fn threshold_decryption_response() {
     let ritual_id = 10;
 
-    let service_secret = RequestSecretKey::random();
+    let service_secret = SessionStaticSecret::random();
 
-    let requester_secret = RequestSecretKey::random();
+    let requester_secret = SessionStaticSecret::random();
     let requester_public_key = requester_secret.public_key();
 
     let decryption_share = b"The Tyranny of Merit";
@@ -782,7 +781,7 @@ fn threshold_decryption_response() {
     );
 
     // wrong secret key used
-    let random_secret_key = RequestSecretKey::random();
+    let random_secret_key = SessionStaticSecret::random();
     let random_shared_secret = random_secret_key.derive_shared_secret(&service_public_key);
     assert!(encrypted_response_from_bytes
         .decrypt(&random_shared_secret)